chore(build/misc): cuda arch 70, sizeof semantics, drop redundant cast
- CMakeLists: include sm_70 in the default CMAKE_CUDA_ARCHITECTURES
  fallback list (70 75 80 86 89)
- model_loader: data_t cudaMalloc bytes written explicitly as
  weightElements() * sizeof(int8_t)
- quantization: drop redundant static_cast<half>(scale)
- TODO.md: mark T1 (full regression + topic commits) done, 184/184 green
- CHANGELOG: record this batch under Unreleased

+++ src/quantization.cpp
@@ -249,7 +249,7 @@ quantizeF16ToW8A16(const half *f16_data, int rows, int cols, int group_size) {
                 scale = 1.0f; // Avoid division by zero for all-zero groups
             }
 
-            scales[static_cast<size_t>(g) * cols + c] = __float2half(static_cast<half>(scale));
+            scales[static_cast<size_t>(g) * cols + c] = __float2half(scale);
 
             // Quantize values
             for (int r = group_start; r < group_end; ++r) {
